+++ pyproject.toml
@@ -105,6 +105,7 @@ wolframalpha = "^5.1.3"
 astra-assistants = "^2.1.0.10"
 composio-langchain = "^0.5.8"
 spider-client = "^0.0.27"
+nltk = "^3.9.1"
 
 
 [tool.poetry.group.dev.dependencies]
