+++ src/backend/base/langflow/main.py
@@ -2,7 +2,6 @@ import asyncio
 import json
 import os
 import warnings
-import nltk
 from contextlib import asynccontextmanager
 from http import HTTPStatus
 from pathlib import Path
@@ -26,6 +25,7 @@ from langflow.initial_setup.setup import (
     create_or_update_starter_projects,
     initialize_super_user_if_needed,
     load_flows_from_directory,
+    download_nltk_resources,
 )
 from langflow.interface.types import get_and_cache_all_types_dict
 from langflow.interface.utils import setup_llm_caching
@@ -78,24 +78,6 @@ class JavaScriptMIMETypeMiddleware(BaseHTTPMiddleware):
         if "files/" not in request.url.path and request.url.path.endswith(".js") and response.status_code == 200:
             response.headers["Content-Type"] = "text/javascript"
         return response
-
-
-# Function to download NLTK packages if not already downloaded
-def download_nltk_resources():
-    nltk_resources = {
-        "corpora": ["wordnet"],
-        "taggers": ["averaged_perceptron_tagger"],
-        "tokenizers": ["punkt", "punkt_tab"],
-    }
-
-    for category, packages in nltk_resources.items():
-        for package in packages:
-            try:
-                nltk.data.find(f"{category}/{package}")
-                logger.info(f"{package} ({category}) already exists.")
-            except LookupError:
-                logger.info(f"Downloading {package} ({category})...")
-                nltk.download(package)
 
 
 def get_lifespan(fix_migration=False, socketio_server=None, version=None):
